+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
 ## Next
-- Fix album grouping
+- Add View menu bar checkbox for album track grouping
+- Use album track grouping when sorting by album, date added, comments, genre, year & artist
 - Fix menu bar checkboxes
 
 ## 0.17.3 - 2024 Apr 30
-- Sub-sort albums by track number when sorting by album
+- Sub-sort albums by track number when sorting by album (album track grouping)
 - Add `Escape` shortcut for de-focusing the sidebar and filter
 - Fix rare error with updating artwork
 - Fix importing of iTunes playlist likes. These aren't visible in the app, but they do get imported. If you previously imported playlists, the likes got lost in Ferrum.